Move LESData to Cases
Move other types outside of src
