+++ driver/Cases.jl
@@ -7,6 +7,7 @@ import OrdinaryDiffEq
 const ODE = OrdinaryDiffEq
 
 import ClimaCore
+import DocStringExtensions
 const CC = ClimaCore
 const CCO = CC.Operators
 const CCG = CC.Geometry
@@ -29,9 +30,6 @@ const TD = Thermodynamics
 import ..TurbulenceConvection
 const TC = TurbulenceConvection
 const TCP = TC.TurbulenceConvectionParameters
-
-import ..TurbulenceConvection: ForcingBase
-import ..TurbulenceConvection: RadiationBase
 
 using ..TurbulenceConvection: pyinterp
 using ..TurbulenceConvection: Grid
@@ -94,11 +92,81 @@ struct DryBubble <: AbstractCaseType end
 struct LES_driven_SCM <: AbstractCaseType end
 
 #####
-##### Case methods
+##### Radiation and forcing types
+#####
+
+struct ForcingNone end
+struct ForcingStandard end
+struct ForcingDYCOMS_RF01 end
+struct ForcingLES end
+
+struct RadiationNone end
+struct RadiationDYCOMS_RF01 end
+struct RadiationLES end
+struct RadiationTRMM_LBA end
+
+"""
+    ForcingBase
+
+LES-driven forcing
+
+$(DocStringExtensions.FIELDS)
+"""
+Base.@kwdef struct ForcingBase{T}
+    "Boolean specifying whether Coriolis forcing is applied"
+    apply_coriolis::Bool = false
+    "Coriolis parameter"
+    coriolis_param::Float64 = 0
+    "Wind relaxation timescale"
+    wind_nudge_τᵣ::Float64 = 0.0
+    "Scalar relaxation lower z"
+    scalar_nudge_zᵢ::Float64 = 0.0
+    "Scalar relaxation upper z"
+    scalar_nudge_zᵣ::Float64 = 0.0
+    "Scalar maximum relaxation timescale"
+    scalar_nudge_τᵣ::Float64 = 0.0
+    "Large-scale divergence (same as in RadiationBase)"
+    divergence::Float64 = 0
+end
+
+ForcingBase(::Type{T}; kwargs...) where {T} = ForcingBase{T}(; kwargs...)
+
+force_type(::ForcingBase{T}) where {T} = T
+
+Base.@kwdef struct RadiationBase{T}
+    "Large-scale divergence (same as in ForcingBase)"
+    divergence::Float64 = 0
+    alpha_z::Float64 = 0
+    kappa::Float64 = 0
+    F0::Float64 = 0
+    F1::Float64 = 0
+end
+
+rad_type(::RadiationBase{T}) where {T} = T
+
+Base.@kwdef struct LESData
+    "Start time index of LES"
+    imin::Int = 0
+    "End time index of LES"
+    imax::Int = 0
+    "Path to LES stats file used to drive SCM"
+    les_filename::String = nothing
+    "Drive SCM with LES data from t = [end - t_interval_from_end_s, end]"
+    t_interval_from_end_s::Float64 = 6 * 3600.0
+    "Length of time to average over for SCM initialization"
+    initial_condition_averaging_window_s::Float64 = 3600.0
+end
+
+#####
+##### Radiation and forcing functions
 #####
 
 include("Radiation.jl")
 include("Forcing.jl")
+
+#####
+##### Case methods
+#####
 
 get_case(namelist::Dict) = get_case(namelist["meta"]["casename"])
 get_case(casename::String) = get_case(Val(Symbol(casename)))
@@ -123,20 +191,20 @@ get_case_name(case_type::AbstractCaseType) = string(case_type)
 ##### Case configurations
 #####
 
-get_forcing_type(::AbstractCaseType) = TC.ForcingStandard # default
-get_forcing_type(::Soares) = TC.ForcingNone
-get_forcing_type(::Nieuwstadt) = TC.ForcingNone
-get_forcing_type(::DYCOMS_RF01) = TC.ForcingDYCOMS_RF01
-get_forcing_type(::DYCOMS_RF02) = TC.ForcingDYCOMS_RF01
-get_forcing_type(::DryBubble) = TC.ForcingNone
-get_forcing_type(::LES_driven_SCM) = TC.ForcingLES
-get_forcing_type(::TRMM_LBA) = TC.ForcingNone
+get_forcing_type(::AbstractCaseType) = ForcingStandard # default
+get_forcing_type(::Soares) = ForcingNone
+get_forcing_type(::Nieuwstadt) = ForcingNone
+get_forcing_type(::DYCOMS_RF01) = ForcingDYCOMS_RF01
+get_forcing_type(::DYCOMS_RF02) = ForcingDYCOMS_RF01
+get_forcing_type(::DryBubble) = ForcingNone
+get_forcing_type(::LES_driven_SCM) = ForcingLES
+get_forcing_type(::TRMM_LBA) = ForcingNone
 
-get_radiation_type(::AbstractCaseType) = TC.RadiationNone # default
-get_radiation_type(::DYCOMS_RF01) = TC.RadiationDYCOMS_RF01
-get_radiation_type(::DYCOMS_RF02) = TC.RadiationDYCOMS_RF01
-get_radiation_type(::LES_driven_SCM) = TC.RadiationLES
-get_radiation_type(::TRMM_LBA) = TC.RadiationTRMM_LBA
+get_radiation_type(::AbstractCaseType) = RadiationNone # default
+get_radiation_type(::DYCOMS_RF01) = RadiationDYCOMS_RF01
+get_radiation_type(::DYCOMS_RF02) = RadiationDYCOMS_RF01
+get_radiation_type(::LES_driven_SCM) = RadiationLES
+get_radiation_type(::TRMM_LBA) = RadiationTRMM_LBA
 
 large_scale_divergence(::Union{DYCOMS_RF01, DYCOMS_RF02}) = 3.75e-6
 
@@ -1112,7 +1180,7 @@ function les_data_kwarg(::LES_driven_SCM, namelist)
         imin = time_interval_bool[1]
         imax = time_interval_bool[end]
 
-        TC.LESData(imin, imax, les_filename, t_interval_from_end_s, namelist["initial_condition_averaging_window_s"])
+        LESData(imin, imax, les_filename, t_interval_from_end_s, namelist["initial_condition_averaging_window_s"])
     end
     return (; LESDat)
 end
