+++ src/EDMF_functions.jl
@@ -10,19 +10,6 @@ function update_cloud_frac(edmf::EDMFModel, grid::Grid, state::State)
     end
 end
 
-function compute_les_Γᵣ(z::FT, τᵣ::FT = 24.0 * 3600.0, zᵢ::FT = 3000.0, zᵣ::FT = 3500.0) where {FT <: Real}
-    # returns height-dependent relaxation timescale from eqn. 9 in `Shen et al. 2021`
-    if z < zᵢ
-        return FT(0)
-    elseif zᵢ <= z <= zᵣ
-        cos_arg = pi * ((z - zᵢ) / (zᵣ - zᵢ))
-        return (FT(0.5) / τᵣ) * (1 - cos(cos_arg))
-    elseif z > zᵣ
-        return (1 / τᵣ)
-    end
-end
-
-
 function compute_turbconv_tendencies!(
     edmf::EDMFModel,
     grid::Grid,
